http and refactoring
