+++ nash-native-client/src/lib.rs
@@ -1,2 +1,6 @@
-pub mod ws_client;
-pub mod http_client;
+pub use types::Environment;
+pub use ws_client::Client;
+
+pub mod http_extension;
+mod types;
+mod ws_client;
