+++ nash-native-client/src/ws_client/absinthe.rs
@@ -85,7 +85,7 @@ impl AbsintheWSResponse {
     }
     pub fn subscription_setup_id(&self) -> Option<String> {
         match &self.payload {
-            Some(ResponsePayload::SubSetup(data)) => Some(data.response.subscription_id.clone()),
+            Some(ResponsePayload::SubscriptionSetup(data)) => Some(data.response.subscription_id.clone()),
             _ => None,
         }
     }
@@ -141,15 +141,16 @@ impl AbsintheInt {
 #[derive(Clone, Deserialize, Debug)]
 #[serde(untagged)]
 pub enum ResponsePayload {
-    SubSetup(SubscriptionSetupResponse), // for subscription setup
+    SubscriptionSetup(SubscriptionSetupResponse), // for subscription setup
     GraphQL(QueryResponse),
     Subscription(SubscriptionResponse),
 }
 
 impl ResponsePayload {
-    pub fn sub_setup(self) -> ProtocolResult<SubscriptionSetupResponse> {
+    #[allow(dead_code)]
+    pub fn subscription_setup(self) -> ProtocolResult<SubscriptionSetupResponse> {
         match self {
-            Self::SubSetup(data) => Ok(data),
+            Self::SubscriptionSetup(data) => Ok(data),
             _ => Err(ProtocolError("Not a subscription setup response")),
         }
     }
