+++ nash-native-client/src/ws_client/stream.rs
@@ -1,10 +1,12 @@
-use super::client::Client;
-use futures::{stream::Stream, StreamExt};
-use nash_protocol::errors::Result;
 use std::{pin::Pin, task::Context, task::Poll};
 
+use futures::{stream::Stream, StreamExt};
+
+use nash_protocol::errors::Result;
 use nash_protocol::protocol::subscriptions::SubscriptionResponse;
 use nash_protocol::protocol::ResponseOrError;
+
+use crate::Client;
 
 impl Stream for Client {
     type Item = Result<ResponseOrError<SubscriptionResponse>>;
